+++ samtranslator/schema/aws_serverless_simpletable.py
@@ -4,7 +4,8 @@ from typing import Optional, Any, Dict
 
 from typing_extensions import Literal
 
-from samtranslator.schema.common import PassThroughProp, BaseModel, get_prop, EmbeddedConnector
+from samtranslator.schema.common import PassThroughProp, BaseModel, get_prop
+from samtranslator.schema.aws_serverless_connector import EmbeddedConnector
 
 primarykey = get_prop("sam-property-simpletable-primarykeyobject")
 properties = get_prop("sam-resource-simpletable")
